+++ lib/joystick/joystick_controller.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
 import 'package:bonfire/base/base_game_point_detector.dart';
-import 'package:bonfire/util/mixins/gestures.dart';
+import 'package:bonfire/util/mixins/pointer_detector_mixin.dart';
 import 'package:bonfire/util/priority_layer.dart';
 import 'package:flame/components/component.dart';
 import 'package:flame/components/mixins/has_game_ref.dart';